--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,4 @@ dist/
 node_modules
 .env
 .idea/
+jsdoc-ast.*