--- a/renderer/.eslintrc
+++ b/renderer/.eslintrc
@@ -13,6 +13,7 @@
         }
     },
 	"env": {
+        "mocha": true,
         "browser": true
     },
     "rules": {