--- a/renderer/src/styles/builtins/customcss.css
+++ b/renderer/src/styles/builtins/customcss.css
@@ -24,7 +24,7 @@
 }
 
 #bd-editor-controls button svg {
-    fill: currentColor;
+    fill: currentcolor;
 }
 
 #bd-editor-controls {