--- a/renderer/src/styles/search.css
+++ b/renderer/src/styles/search.css
@@ -17,7 +17,7 @@
     flex: 1;
 }
 
-.bd-search::-webkit-input-placeholder {
+.bd-search::input-placeholder {
     color: var(--text-muted);
 }