--- a/renderer/src/styles/ui/changelog.css
+++ b/renderer/src/styles/ui/changelog.css
@@ -107,7 +107,7 @@
 }
 
 .bd-changelog-modal .bd-modal-content a {
-    color: hsl(200, calc(var(--saturation-factor, 1) * 100%), 49.4%);
+    color: hsl(200deg calc(var(--saturation-factor, 1) * 100%) 49.4%);
     transition: 0.05s;
     text-decoration: none;
 }
@@ -119,21 +119,21 @@
 .theme-dark .bd-changelog-modal .bd-modal-content ol,
 .theme-dark .bd-changelog-modal .bd-modal-content p,
 .theme-dark .bd-changelog-modal .bd-modal-content ul li {
-    color: hsl(210, calc(var(--saturation-factor, 1) * 9.3%), 78.8%);
+    color: hsl(210deg calc(var(--saturation-factor, 1) * 9.3%) 78.8%);
 }
 
 .theme-dark .bd-changelog-modal .bd-modal-content ul li::before {
-    background-color: hsl(216, calc(var(--saturation-factor, 1) * 9.8%), 90%);
+    background-color: hsl(216deg calc(var(--saturation-factor, 1) * 9.8%) 90%);
 }
 
 .theme-light .bd-changelog-modal .bd-modal-content ol,
 .theme-light .bd-changelog-modal .bd-modal-content p,
 .theme-light .bd-changelog-modal .bd-modal-content ul li {
-    color: hsl(223, calc(var(--saturation-factor, 1) * 5.8%), 52.9%);
+    color: hsl(223deg calc(var(--saturation-factor, 1) * 5.8%) 52.9%);
 }
 
 .theme-light .bd-changelog-modal .bd-modal-content ul li::before {
-    background-color: hsl(223, calc(var(--saturation-factor, 1) * 5.8%), 52.9%);
+    background-color: hsl(223deg calc(var(--saturation-factor, 1) * 5.8%) 52.9%);
 }
 
 .bd-changelog-title {
@@ -158,7 +158,7 @@
     height: 1px;
     flex: 1 1 auto;
     margin-left: 4px;
-    opacity: .6;
+    opacity: 0.6;
 }
 
 .bd-changelog-added {
@@ -170,11 +170,11 @@
 }
 
 .bd-changelog-fixed {
-    color: hsl(359, calc(var(--saturation-factor, 1)*87.3%), 59.8%);
+    color: hsl(359deg calc(var(--saturation-factor, 1) *87.3%) 59.8%);
 }
 
 .bd-changelog-fixed::after {
-    background-color: hsl(359, calc(var(--saturation-factor, 1)*87.3%), 59.8%);
+    background-color: hsl(359deg calc(var(--saturation-factor, 1) *87.3%) 59.8%);
 }
 
 .bd-changelog-progress {
@@ -186,31 +186,31 @@
 }
 
 .bd-changelog-improved {
-    color: hsl(235, calc(var(--saturation-factor, 1)*85.6%), 64.7%);
+    color: hsl(235deg calc(var(--saturation-factor, 1) *85.6%) 64.7%);
 }
 
 .bd-changelog-improved::after {
-    background-color: hsl(235, calc(var(--saturation-factor, 1)*85.6%), 64.7%);
+    background-color: hsl(235deg calc(var(--saturation-factor, 1) *85.6%) 64.7%);
 }
 
 .theme-dark .bd-changelog-improved {
-    color: hsl(235, calc(var(--saturation-factor, 1)*86.1%), 77.5%);
+    color: hsl(235deg calc(var(--saturation-factor, 1) *86.1%) 77.5%);
 }
 
 .theme-dark .bd-changelog-improved::after {
-    background-color: hsl(235, calc(var(--saturation-factor, 1)*86.1%), 77.5%);
+    background-color: hsl(235deg calc(var(--saturation-factor, 1) *86.1%) 77.5%);
 }
 
 .theme-dark .bd-changelog-modal iframe,
 .theme-dark .bd-changelog-modal video,
 .theme-dark .bd-changelog-modal img {
-    box-shadow: 0 2px 10px 0 hsl(var(0, calc(var(--saturation-factor, 1)*0%), 0%-hsl)/.2);
+    box-shadow: 0 2px 10px 0 hsl(var(0, calc(var(--saturation-factor, 1) *0%), 0%-hsl)/20%);
 }
 
 .theme-light .bd-changelog-modal iframe,
 .theme-light .bd-changelog-modal video,
 .theme-light .bd-changelog-modal img {
-    box-shadow: 0 2px 10px 0 hsl(var(0, calc(var(--saturation-factor, 1)*0%), 0%-hsl)/.1);
+    box-shadow: 0 2px 10px 0 hsl(var(0, calc(var(--saturation-factor, 1) *0%), 0%-hsl)/10%);
 }
 
 /* .socialLink-1qjJIk {