Since breaking changes planned (development branch), this PR can be useful only for patches (v1.12.x).

## Features
- Make scripts faster and shorter by using a single command instead of a command stack: `webpack && webpack && webpack` → `webpack`. This also results in shorter output. Prefer using configs over multiple runs.
![image](https://github.com/user-attachments/assets/435ea4c9-7058-4b0c-85e0-699d5cce5a61)
- Add more commands for sub-packages.
- Run `pnpm lint` across the entire workspace. It should work for every sub-package and the root (common/**.js).
![image](https://github.com/user-attachments/assets/926990c5-5581-4d00-b54b-bcc9eeb3f3aa)

## Other improvements
- Add `jsdoc@4` as a dependency and ignore the output of `pnpm run api` by adding it to `.gitignore`.
- Use the Mocha environment in `.eslintrc` and add Mocha types as a dev dependency.

> [!NOTE]
> `pnpm lint` causes errors because of the stylelint. JavaScript is linted fine.
>
> Both CSS and JavaScript are linted, which is why there are so many changes in the CSS files.