@Mopsgamer 2025-01-29
### Benchmarks ⚡
#### `pnpm run dist`
`hyperfine -m 3 --warmup 2 --prepare 'git checkout improve-workspace-scripts && pnpm i && pnpm i -r' -n 'new pnpm run dist' 'pnpm run dist' --prepare 'git checkout main && pnpm i && pnpm i -r' -n 'old pnpm run dist' 'pnpm run dist'`
```
Benchmark 1: new pnpm run dist
  Time (mean ± σ):     17.351 s ±  0.141 s    [User: 20.844 s, System: 2.880 s]
  Range (min … max):   17.254 s … 17.513 s    3 runs
 
Benchmark 2: old pnpm run dist
  Time (mean ± σ):     19.595 s ±  0.230 s    [User: 16.062 s, System: 1.677 s]
  Range (min … max):   19.354 s … 19.813 s    3 runs
 
Summary
  new pnpm run dist ran
    1.13 ± 0.02 times faster than old pnpm run dist
```

#### `pnpm run build`
`hyperfine -m 3 --warmup 2 --prepare 'git checkout improve-workspace-scripts && pnpm i && pnpm i -r' -n 'new pnpm run build' 'pnpm run build' --prepare 'git checkout main && pnpm i && pnpm i -r' -n 'old pnpm run build' 'pnpm run build'`
```
Benchmark 1: new pnpm run build
  Time (mean ± σ):     11.274 s ±  0.964 s    [User: 12.932 s, System: 1.952 s]
  Range (min … max):   10.172 s … 11.963 s    3 runs
 
Benchmark 2: old pnpm run build
  Time (mean ± σ):     15.640 s ±  0.074 s    [User: 10.604 s, System: 1.566 s]
  Range (min … max):   15.565 s … 15.712 s    3 runs
 
Summary
  new pnpm run build ran
    1.39 ± 0.12 times faster than old pnpm run build
```

#### `pnpm run build-prod`
`hyperfine -m 3 --warmup 2 --prepare 'git checkout improve-workspace-scripts && pnpm i && pnpm i -r' -n 'new pnpm build-prod' 'pnpm run build-prod' --prepare 'git checkout main && pnpm i && pnpm i -r' -n 'old pnpm build-prod' 'pnpm run build-prod'`
```
Benchmark 1: new pnpm build-prod
  Time (mean ± σ):     27.053 s ±  1.328 s    [User: 30.484 s, System: 2.971 s]
  Range (min … max):   25.544 s … 28.042 s    3 runs
 
Benchmark 2: old pnpm build-prod
  Time (mean ± σ):     21.230 s ±  1.551 s    [User: 16.192 s, System: 1.721 s]
  Range (min … max):   19.753 s … 22.846 s    3 runs
 
Summary
  old pnpm build-prod ran
    1.27 ± 0.11 times faster than new pnpm build-prod
```