@zerebos 2025-02-02
Hi I appreciate the work that went into that but we're about to go through some major shifts in how we organize the repository including moving off of `webpack` entirely.