@Mopsgamer 2025-02-02
Understood. Since the transition will take time, this PR could still help speed up development of patches in the meantime. But if you think it might cause issues, feel free to close it.